requirements.txt
appdirs==1.4.0
appnope==0.1.0
arrow==0.8.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
binaryornot==0.4.0
blessings==1.6
boto==2.39.0
certifi==2016.2.28
cffi==1.6.0
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
coverage==4.0.3
decorator==4.0.10
dj-database-url==0.4.0
Django==1.9.7
django-activity-stream==0.6.2
django-allauth==0.25.2
django-bootstrap3==7.1.0
django-braces==1.9.0
django-configurations==1.0
django-cors-headers==1.1.0
django-extensions==1.6.7
django-filter==0.13.0
django-hstore==1.4.2
django-location-field==1.6.1
django-model-utils==2.4
django-mptt==0.8.4
django-nose==1.4.3
django-oauth-toolkit==0.10.0
django-redis-cache==1.6.5
django-rest-auth==0.8.0
django-rest-framework-social-oauth2==1.0.4
django-rq==0.9.0
django-secure==1.0.1
django-storages-redux==1.2.3
django-tagulous==0.11.1
django-unique-upload==0.2.1
django-vanilla-views==1.0.4
django-versatileimagefield==1.3
djangorestframework==3.3.3
djangorestframework-filters==0.8.0
djangorestframework-gis==0.10.1
djangorestframework-hstore==1.3
djangorestframework-jwt==1.8.0
enum34==1.1.6
Fabric==1.10.2
factory-boy==2.6.1
fake-factory==0.5.7
flake8==2.5.4
funcsigs==1.0.2
future==0.15.2
gevent==1.0.2
gnureadline==6.3.3
greenlet==0.4.10
gunicorn==19.3.0
hiredis==0.2.0
idna==2.1
ipaddress==1.0.16
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
Jinja2==2.8
jinja2-time==0.2.0
livereload==2.4.1
Markdown==2.6.6
MarkupSafe==0.23
mccabe==0.4.0
mkdocs==0.15.3
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0
mock==1.3.0
newrelic==2.50.0.39
nose==1.3.7
nose-progressive==1.5.1
oauthlib==1.0.3
packaging==16.8
paramiko==2.0.0
pathlib2==2.1.0
pbr==1.10.0
pep8==1.7.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
poyo==0.4.0
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
pyflakes==1.0.0
PyJWT==1.4.0
pyparsing==2.1.10
python-dateutil==2.5.3
python-openid==2.2.5
python-social-auth==0.2.19
python-twitter==3.2
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
rq==0.6.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
traitlets==4.2.1
urllib3==1.16
vanilla==0.1.16
Werkzeug==0.11.10
whichcraft==0.4.0
whitenoise==3.0
